noticias_python/requirements.txt
aniso8601==8.0.0
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
flask-mongoengine==0.9.5
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
python-dotenv==0.14.0
pytz==2020.1
shell==1.0.1
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
